test(rust): improve `TCP` tests
